Add buttons to the floating tool window. They don't do anything yet apart from look pretty. Give some order to the resource.h file tool.
svn path=/trunk/; revision=21552

+++ reactos/base/applications/imagesoft/imagesoft.c
@@ -32,7 +32,7 @@ WinMain(HINSTANCE hThisInstance,
         return 1;
     }
 
-    len = _tcslen(lpAppName) + _tcslen(lpVersion);
+    len = (int)_tcslen(lpAppName) + (int)_tcslen(lpVersion);
     lpTitle = HeapAlloc(ProcessHeap,
                         0,
                         (len + 2) * sizeof(TCHAR));
